requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-core
armoriq-sdk
httpx
pydantic
